+++ CHANGELOG.md
@@ -12,6 +12,8 @@ This project **DOES NOT** adhere to [Semantic Versioning](http://semver.org/spec
 - Keep `close` related contract in `DualConnection` API
 - Make `close` safe
 - Release `Statement` when closed
+- Keep `close` related contract in `Statement` API
+- Keep `close` related contract in `PreparedStatement` API
 
 ## [0.1.12] - 2020-11-27
 [0.1.12]: https://bitbucket.org/atlassian/db-replica/branches/compare/release-0.1.12%0Drelease-0.1.11